Clean up wsgi
