+++ App/__init__.py
@@ -1,4 +1,5 @@
 from .models import *
 from .views import *
 from .controllers import *
+from .utils import *
 from .main import *
